app/requirements.txt
pywebio==1.8.3
pywebio-battery==0.7.0
loguru==0.7.2
pytest==8.3.2
fastapi==0.112.2
uvicorn==0.30.6
websockets==13.0.1
python-dotenv==1.0.1
prisma==0.13.1
minio==7.2.7